+++ delivery-service/src/main/java/com/fn/eureka/client/deliveryservice/config/HubRouteScheduler.java
@@ -4,7 +4,7 @@ import org.springframework.scheduling.annotation.EnableScheduling;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-import com.fn.eureka.client.deliveryservice.application.HubToHubService;
+import com.fn.eureka.client.deliveryservice.domain.service.HubToHubServiceImpl;
 
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
@@ -15,14 +15,14 @@ import lombok.extern.slf4j.Slf4j;
 @EnableScheduling
 public class HubRouteScheduler {
 
-	private final HubToHubService hubToHubService;
+	private final HubToHubServiceImpl hubToHubServiceImpl;
 
 	@Scheduled(cron = "0 30 * * * *")
 	public void updateHubRoutes() {
 
 		log.info("주기적인 허브 루트 갱신");
 
-		hubToHubService.updateAllRoutes();
+		hubToHubServiceImpl.updateAllRoutes();
 
 		log.info("허브 업데이트 종료");
